Make Filter workable (Combination Filter)
We use groovy script strategy to do not run
build targets for absent configs (Valhalla project). At the
same time we pass a filter in groovy format as a parameter
to ENG jobs. Then we apply this groovy as Combination Filter
but it can be used within Classic execution strategy only.
So I merged part of Classic strategy here in Groovy
Strategy. And now we are triggering configuration if it
satisfy both filters: Combination and Groovy Strategy.

task:GENCM2018-423