Run script got from file inside sandbox.
We need it to disable security groovy script approval for
scripts run from file.
Jenkins plugin permissive-script-security-plugin allows us to do this.
But it disables approvals only for whitelist based mechanism which is
taking place for groovy sandboxes only.

Folow the instruction on page
https://plugins.jenkins.io/permissive-script-security/
to activate the plugin.

task:GENCM2018-439